Move some tests to ReferenceFinderITest.
