+++ plugin/src/test/java/io/jenkins/plugins/analysis/warnings/steps/StepsITest.java
@@ -862,97 +862,6 @@ public class StepsITest extends IntegrationTestWithJenkinsPerTest {
     }
 
     /**
-     * Creates a reference job and starts the analysis for this job. Then another job is created that uses the first one
-     * as reference. Verifies that the association is correctly stored.
-     */
-    @Test
-    public void shouldUseOtherJobAsReference() {
-        WorkflowJob reference = createPipeline("reference");
-        copyMultipleFilesToWorkspaceWithSuffix(reference, "java-start.txt");
-        reference.setDefinition(createPipelineScriptWithScanAndPublishSteps(new Java()));
-
-        AnalysisResult referenceResult = scheduleSuccessfulBuild(reference);
-
-        assertThat(referenceResult.getTotalSize()).isEqualTo(2);
-        assertThat(referenceResult.getIssues()).hasSize(2);
-        assertThat(referenceResult.getReferenceBuild()).isEmpty();
-
-        WorkflowJob job = createPipelineWithWorkspaceFiles("java-start.txt");
-        job.setDefinition(asStage(createScanForIssuesStep(new Java()),
-                "publishIssues issues:[issues], referenceJobName:'reference'"));
-
-        AnalysisResult result = scheduleSuccessfulBuild(reference);
-
-        assertThat(result.getTotalSize()).isEqualTo(2);
-        assertThat(result.getIssues()).hasSize(2);
-        assertThat(result.getReferenceBuild()).hasValue(referenceResult.getOwner());
-
-        // TODO: add verification for io.jenkins.plugins.analysis.core.model.IssueDifference
-    }
-
-    /**
-     * Creates a reference job and starts a build having 2 warnings.  Then two builds for the job.  Then another job is created that
-     * uses the first build as a reference.  Verifies that the association is correctly stored.
-     */
-    @Test
-    public void shouldUseOtherJobBuildAsReference() {
-        WorkflowJob reference = createPipeline("reference");
-        copyMultipleFilesToWorkspaceWithSuffix(reference, "java-start-rev0.txt");
-        reference.setDefinition(createPipelineScriptWithScanAndPublishSteps(new Java()));
-
-        AnalysisResult firstReferenceResult = scheduleSuccessfulBuild(reference);
-        cleanWorkspace(reference);
-        copyMultipleFilesToWorkspaceWithSuffix(reference, "java-start.txt");
-        AnalysisResult secondReferenceResult = scheduleSuccessfulBuild(reference);
-
-        assertThat(firstReferenceResult.getTotalSize()).isEqualTo(1);
-        assertThat(firstReferenceResult.getIssues()).hasSize(1);
-        assertThat(firstReferenceResult.getReferenceBuild()).isEmpty();
-        assertThat(firstReferenceResult.getOwner().getId()).isEqualTo("1");
-
-        assertThat(secondReferenceResult.getTotalSize()).isEqualTo(2);
-        assertThat(secondReferenceResult.getIssues()).hasSize(2);
-        assertThat(secondReferenceResult.getReferenceBuild().get().getId()).isEqualTo("1");
-        assertThat(secondReferenceResult.getOwner().getId()).isEqualTo("2");
-
-        WorkflowJob job = createPipelineWithWorkspaceFiles("java-start.txt");
-        job.setDefinition(asStage(createScanForIssuesStep(new Java()),
-                "publishIssues issues:[issues], referenceJobName:'reference', referenceBuildId: '1'"));
-
-        AnalysisResult result = scheduleSuccessfulBuild(job);
-
-        assertThat(result.getReferenceBuild().isPresent());
-        assertThat(result.getReferenceBuild().get().getId()).isEqualTo(firstReferenceResult.getOwner().getId());
-        assertThat(result.getReferenceBuild().get().getId()).isEqualTo("1");
-
-        assertThat(result.getNewIssues()).hasSize(1);
-    }
-
-    /**
-     * Creates a reference job without builds, then builds the job, referring to a non-existant build
-     * in the reference job.
-     */
-    @Test
-    public void shouldHandleMissingJobBuildAsReference() {
-        WorkflowJob reference = createPipeline("reference");
-        copyMultipleFilesToWorkspaceWithSuffix(reference, "java-start-rev0.txt");
-        reference.setDefinition(createPipelineScriptWithScanAndPublishSteps(new Java()));
-
-        WorkflowJob job = createPipelineWithWorkspaceFiles("java-start.txt");
-        job.setDefinition(asStage(createScanForIssuesStep(new Java()),
-                "publishIssues issues:[issues], referenceJobName:'reference'"));
-
-        AnalysisResult result = scheduleSuccessfulBuild(job);
-
-        assertThat(not(result.getReferenceBuild().isPresent()));
-
-        assertThat(result.getNewIssues()).hasSize(0);
-        assertThat(result.getOutstandingIssues()).hasSize(2);
-        assertThat(result.getErrorMessages()).contains(
-                "Reference job 'reference' does not contain any valid build");
-    }
-
-    /**
      * Creates a reference job with a build, then builds the job, referring to a non-existing build
      * in the reference job.
      */
